double

+++ LibraryProject/src/Vector2/Vector2.cpp
@@ -1,6 +1,6 @@
 #include "Vector2.h"
 
-Vector2::Vector2(float x, float y)
+Vector2::Vector2(double x, double y)
 {
 	this->x = x;
 	this->y = y;
